pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "google-generativeai>=0.8.4",
    "langchain-google-genai>=2.0.10",
    "langchain>=0.3.23",
    "langchain-openai>=0.3.12",
    "python-dotenv>=1.1.0",
]
